requirements.txt
protobuf==6.30.1
python-dotenv==1.0.1
streamlit==1.34.0
google-generativeai==0.8.4
